requirements.txt
annotated-types==0.7.0
anyio==4.5.2
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.13.4
black==21.9b0
bleach==6.1.0
blinker==1.8.2
Brlapi==0.7.0
certifi==2025.8.3
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.3
click==7.1.2
cloudinary==1.44.1
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
distro-info===0.23ubuntu1
dj-database-url==0.4.1
Django==4.2.25
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.10
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.3.0
executing==2.2.0
fasteners==0.14.1
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
fonttools==4.57.0
future==0.18.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.14.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.1
json5==0.12.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.1.3
monotonic==1.5
mypy_extensions==1.1.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.4
notebook_shim==0.2.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
openai==1.93.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==3.6.1
psutil==7.0.0
psycopg==3.2.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.4
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==3.3.0
pytokens==0.1.10
pytz==2025.2
pyxdg==0.26
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.35.1
regex==2024.11.6
reportlab==3.5.34
requests==2.32.4
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
seaborn==0.13.2
SecretStorage==2.3.1
selenium==4.27.1
Send2Trash==1.8.3
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
systemd-python==234
terminado==0.18.1
timedelta==2020.12.3
tinycss2==1.2.1
tomli==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.3
usb-creator==0.3.7
virtualenv==20.31.2
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
whitenoise==6.7.0
wsproto==1.2.0
xkit==0.0.0
zipp==3.20.2
